various compile fixes
git-svn-id: svn://svn.dd-wrt.com/DD-WRT@31606 52c4871e-980c-0410-b1e0-e73912ce01f8

Former-commit-id: a07ed28352dfa3a1e9a25a4797295611e18f0c19

+++ src/linux/universal/linux-4.9/drivers/net/ethernet/intel/e1000gbe/iegbe_osdep.h
@@ -90,7 +90,7 @@ typedef enum {
 
 
 #define E1000_WRITE_REG(a, reg, value) ( \
-    writel((value), ((a)->hw_addr + \
+    writel((unsigned int)(value), ((a)->hw_addr + \
         (((a)->mac_type >= iegbe_82543) ? E1000_##reg : E1000_82542_##reg))))
 
 #define E1000_READ_REG(a, reg) ( \
